Move QuestionResultLoader beside its only consumer, NotebookStepPreview
