+++ frontend/src/metabase/querying/notebook/components/NotebookStep/NotebookStepPreview/QuestionResultLoader.unit.spec.tsx
